+++ src/main/java/kr/hhplus/be/server/domain/order/OrderService.java
@@ -3,8 +3,10 @@ package kr.hhplus.be.server.domain.order;
 import kr.hhplus.be.server.common.exception.BusinessError;
 import kr.hhplus.be.server.common.exception.BusinessException;
 import lombok.RequiredArgsConstructor;
+import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.transaction.support.TransactionTemplate;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -18,6 +20,7 @@ public class OrderService {
     private final OrderRepository orderRepository;
     private final OrderValidators orderValidators;
     private final OrderExternalClient orderExternalClient;
+    private final TransactionTemplate transactionTemplate;
 
     /**
      * 주문 ID를 기반으로 주문 내역을 조회합니다.
@@ -160,5 +163,16 @@ public class OrderService {
         orderRepository.saveOrder(order);
     }
 
+    @Scheduled(fixedRate = 120000)
+    public void expireOrders() {
+        LocalDateTime expirableDateTime = LocalDateTime.now().minusMinutes(5);
+
+        List<Order> expiredOrders = orderRepository.findExpiredOrders(expirableDateTime);
+
+        expiredOrders.forEach(order -> transactionTemplate.execute(status -> {
+            order.cancel();
+            return orderRepository.saveOrder(order);
+        }));
+    }
 
 }
